# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NovaDev Cloud IDE API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStatsLanguageBreakdownItem } from './projectStatsLanguageBreakdownItem';

export interface ProjectStats {
  totalProjects: number;
  recentlyActive: number;
  languageBreakdown: ProjectStatsLanguageBreakdownItem[];
  totalDeployments: number;
  storageUsedMb?: number;
}